+++ src/instruments/instrument_ext.h
@@ -1,5 +1,5 @@
-#ifndef INSTRUMENT_SENO
-#define INSTRUMENT_SENO
+#ifndef INSTRUMENT_EXT
+#define INSTRUMENT_EXT
 
 #include <vector>
 #include <string>
@@ -7,13 +7,13 @@
 #include "envelope_adsr.h"
 
 namespace upc {
-  class InstrumentSeno: public upc::Instrument {
+  class InstrumentExt: public upc::Instrument {
     EnvelopeADSR adsr;
     float phase, step;
 	float A;
     std::vector<float> tbl;
   public:
-    InstrumentSeno(const std::string &param = "");
+    InstrumentExt(const std::string &param = "");
     void command(long cmd, long note, long velocity=1); 
     const std::vector<float> & synthesize();
     bool is_active() const {return bActive;}
