+++ work/dumb.orc
@@ -1,4 +1,7 @@
 1	InstrumentDumb	ADSR_A=0.02; ADSR_D=0.1; ADSR_S=0.4; ADSR_R=0.1; N=40;
-2	InstrumentDumb	ADSR_A=0.01; ADSR_D=0.1; ADSR_S=0; ADSR_R=0; N=40; // Percusión, mantiene la nota pulsada hasta su completa extinción
-3	InstrumentDumb	ADSR_A=0.01; ADSR_D=0.5; ADSR_S=0; ADSR_R=0; N=40; // Percusión, finaliza la nota antes de su completa extinción
-4	InstrumentDumb	ADSR_A=0.1; ADSR_D=0; ADSR_S=0.8; ADSR_R=0.05; N=40; // Flauta
+2	InstrumentDumb	ADSR_A=0.01; ADSR_D=0.1; ADSR_S=0; ADSR_R=0; N=40; # Percusión, mantiene la nota pulsada hasta su completa extinción
+3	InstrumentDumb	ADSR_A=0.01; ADSR_D=0.5; ADSR_S=0; ADSR_R=0; N=40; # Percusión, finaliza la nota antes de su completa extinción
+4	InstrumentDumb	ADSR_A=0.1; ADSR_D=0; ADSR_S=0.8; ADSR_R=0.05; N=40; # Flauta
+5	InstrumentExt	ADSR_A=0.1; ADSR_D=0; ADSR_S=0.8; ADSR_R=0.05; N 0 40; File = "file"; # Tabla Exterior
+6	InstrumentSamp	ADSR_A=0.1; ADSR_D=0; ADSR_S=0.8; ADSR_R=0.05; N 0 40; # Sampler
+
